refactor: getMultichainBlockExplorer to be a hook
