+++ app/components/Views/AccountActions/AccountActions.tsx
@@ -17,10 +17,6 @@ import BottomSheet, {
 } from '../../../component-library/components/BottomSheets/BottomSheet';
 import AccountAction from '../AccountAction/AccountAction';
 import { IconName } from '../../../component-library/components/Icons/Icon';
-import {
-  findBlockExplorerForNonEvmAccount,
-  getBlockExplorerName,
-} from '../../../util/networks';
 
 import { MetaMetricsEvents } from '../../../core/Analytics';
 import { selectProviderConfig } from '../../../selectors/networkController';
@@ -50,6 +46,7 @@ import { useTheme } from '../../../util/theme';
 import { useEIP7702Networks } from '../confirmations/hooks/7702/useEIP7702Networks';
 import { isEvmAccountType } from '@metamask/keyring-api';
 import { toHex } from '@metamask/controller-utils';
+import { getMultichainBlockExplorer } from '../../../core/Multichain/networks';
 
 interface AccountActionsParams {
   selectedAccount: InternalAccount;
@@ -91,28 +88,10 @@ const AccountActions = () => {
         title: string;
         blockExplorerName: string;
       }
-    | undefined = useMemo(() => {
-    if (selectedAccount) {
-      if (isEvmAccountType(selectedAccount.type)) {
-        return {
-          url: `https://etherscan.io/address/${selectedAccount.address}#asset-multichain`,
-          title: 'Etherscan (Multichain)',
-          blockExplorerName: 'Etherscan (Multichain)',
-        };
-      }
-
-      const explorer = findBlockExplorerForNonEvmAccount(selectedAccount);
-      if (explorer) {
-        return {
-          url: explorer,
-          title: new URL(explorer).hostname,
-          blockExplorerName:
-            getBlockExplorerName(explorer) ?? new URL(explorer).hostname,
-        };
-      }
-      return undefined;
-    }
-  }, [selectedAccount]);
+    | undefined = useMemo(
+    () => getMultichainBlockExplorer(selectedAccount),
+    [selectedAccount],
+  );
 
   const goToBrowserUrl = (url: string, title: string) => {
     navigate('Webview', {
